*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@133 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/cspfj/MCRW.java
@@ -123,7 +123,7 @@ public final class MCRW extends AbstractLocalSolver {
 
 		init();
 
-		int nbConflicts = weightedConflicts();
+		int nbConflicts = (int)weightedConflicts();
 
 		logger.fine("Searching...");
